requirements.txt
yfinance
fastapi
supabase
python-dotenv